+++ CHANGELOG.md
@@ -2,9 +2,9 @@
 
 ## Unreleased
 
-## v0.3.1 - 2024/12/21
-
 - Fix: dropping `Print<Background<W>>` now emits a message of `(Error)` and a newline (https://github.com/heroku-buildpacks/bullet_stream/pull/20)
+- Added: `bullet_stream::global::print` functions for writing formatted output without needing to preserve state (https://github.com/heroku-buildpacks/bullet_stream/pull/21)
+- Added: `Print::global()` and `bullet_stream::global::set_writer`. Use these to preserve the newline indentation when handling dropped structs or errors (https://github.com/heroku-buildpacks/bullet_stream/pull/21)
 
 ## v0.3.0 - 2024/08/14